Improve payment adding interface
