+++ client/src/global-styles.scss
@@ -14,6 +14,22 @@ body {
 	background-color: #057e36;
 }
 
+table {
+  border-collapse: collapse;
+  border: 1px solid;
+  td, th {
+    padding: 0.2rem 0.4rem;
+    border: 1px solid;
+  }
+}
+
+.button-container {
+  margin-top: 1rem;
+  & > * {
+    margin-right: 1rem;
+  }
+}
+
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
 }
