+++ client/src/pages/EventAdminRegistrations/EditTab/PaymentTab.tsx
@@ -4,76 +4,48 @@ import {
   Button,
 } from 'react-bootstrap';
 import Spinner from 'components/Spinner';
-import Modal from 'components/Modal';
-import JsonSchemaForm from 'components/JsonSchemaForm';
-import Input, { Money, Select } from 'components/Input';
-import { formatDateTimeForApi, formatDateTimeForViewing } from 'utils/time';
+import {
+  formatDateTimeForViewing,
+} from 'utils/time';
 import getFromPath from 'lodash/get';
 
 import api from 'hooks/api';
+
+import AddPaymentFormModal from './AddPaymentFormModal';
 
 interface Props {
   event: ApiEvent;
   registration: AugmentedRegistration;
 }
 
+
 function PaymentTab(props: Props) {
   const paymentApi = api.useGetPaymentsQuery();
-  const [createPayment, createPaymentResult] = api.useCreatePaymentMutation();
-  const modalRef = React.useRef<Modal>(null);
-  const [paymentAttributes, setPaymentAttributes] = React.useState({});
-  const [paymentAmount, setPaymentAmount] = React.useState(0);
-  const [paymentDate, setPaymentDate] = React.useState('');
-  const [paymentType, setPaymentType] = React.useState('Check');
+  const [showAddPayment, setShowAddPayment] = React.useState(false);
 
   if (paymentApi.isLoading || !paymentApi.data) return <Spinner />;
-  if (createPaymentResult.isLoading) return <Spinner />;
 
   const payments = paymentApi.data.filter(
     p => p.registration === props.registration.id
   );
 
-  const handleChange = (setter: Function) => (changeEvent: React.ChangeEvent<HTMLInputElement>) => {
-    let { value } = changeEvent.target;
-
-    if (!value) return;
-
-    setter(value);
-  };
-
-  const showAddPaymentModal = () => modalRef.current && modalRef.current.show();
-  const handleSaveNewPayment = () => {
-    const dateValue = formatDateTimeForApi(paymentDate) || null;
-    const data = {
-      registration: props.registration.id,
-      payment_type: paymentType,
-      paid_on: dateValue,
-      attributes: paymentAttributes,
-      deposit: null,
-      amount: Number(paymentAmount) || 0,
-    };
-
-    try {
-      createPayment(data);
-    } catch (e) {
-      console.error(e);
-    }
-
-    setPaymentAttributes({});
-    setPaymentAmount(0);
-    setPaymentDate('');
-  };
-
-  // # from payments
-  // paid_on = models.DateTimeField(null=True)
-  // attributes = models.JSONField(null=True)
-  // amount = models.DecimalField(max_digits=7, decimal_places=2, default=Decimal('0.00'))
+  const showAddPaymentModal = () => setShowAddPayment(true);
 
   const paymentSchema = getFromPath(props, 'event.payment_schema.properties', { properties: {} });
-  const keys = Object.keys(paymentSchema).sort();
+  const keys = Object.keys(paymentSchema.properties).sort();
 
   return (
     <div>
+      <p>
+        <div>
+          Payment Type at Registration: {
+            props.registration.payment_type || 'None'
+          }
+        </div>
+        <div>Total Owed: ${props.registration.total_owed}</div>
+        <div>Total Payments: ${props.registration.total_payments}</div>
+        <div>Balance Due: ${props.registration.total_balance}</div>
+      </p>
       {
         payments.length <= 0 ? (
           <p>
@@ -112,47 +84,16 @@ function PaymentTab(props: Props) {
           </table>
         )
       }
-      <p>
+      <p className="button-container">
         <Button onClick={showAddPaymentModal}>Add Payment</Button>
       </p>
-      <Modal
-        ref={modalRef}
-        title="Add Payment"
-        saveButtonLabel="Add new payment"
-        onSave={handleSaveNewPayment}
-      >
-        <div>
-            <Select
-              label="Type"
-              options={[
-                'Check',
-                'PayPal',
-              ].map(v => ({ label: v, value: v }))}
-              onChange={handleChange(setPaymentType)}
-            />
-            <Input
-              label="Paid date"
-              onChange={handleChange(setPaymentDate)}
-              type="date"
-              defaultValue={paymentDate}
-            />
-            <Money
-              label="Amount"
-              onChange={handleChange(setPaymentAmount)}
-              defaultValue={paymentAmount}
-            />
-            <JsonSchemaForm
-              schema={paymentSchema}
-              onChange={(f) => setPaymentAttributes(f.formData)}
-              onSubmit={handleSaveNewPayment}
-              onError={() => console.log("errors")}
-              formData={paymentAttributes}
-              templateData={{}}
-            >
-              <span />
-            </JsonSchemaForm>
-        </div>
-      </Modal>
+      <AddPaymentFormModal
+        event={props.event}
+        registration={props.registration}
+        show={showAddPayment}
+        onClose={() => setShowAddPayment(false)}
+        onSave={() => setShowAddPayment(false)}
+      />
     </div>
   );
 }
